feat: new file structure
