+++ db/db.go
@@ -1,0 +1,24 @@
+package db
+
+import (
+	"database/sql"
+	"log"
+
+	_ "github.com/mattn/go-sqlite3"
+)
+
+func InitializeDB(dbFile string) (*sql.DB, error) {
+	db, err := sql.Open("sqlite3", dbFile)
+	if err != nil {
+		log.Fatalf("Error opening database: %v", err)
+		return nil, err
+	}
+
+	err = db.Ping()
+	if err != nil {
+		log.Fatalf("Error connecting to the database: %v", err)
+		return nil, err
+	}
+
+	return db, nil
+}
